Removing "Not Applicable*" from EoM and EoS as it is confusing
This makes is explicit for when a given release is going to be out
of maintenance and support and prevents readers from having to hunt
and peck around the table.

Of particular note, LA release EoM and EoS dates should be 6 months
and one year from the GA date. This is very confusing to people, so
I have decided provide a date based on release and those dates must
be updated once the software is GAed